SAMSON-291 Information about size manager by an Endpoint in 64 bits
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@2204 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/samson/network/Endpoint2.h
@@ -142,8 +142,8 @@ public:
 	// Statistics
 	int                  msgsIn;
 	int                  msgsOut;
-	int                  bytesIn;
-	int                  bytesOut;
+	size_t               bytesIn;
+	size_t               bytesOut;
 	int                  msgsInErrors;
 	int                  msgsOutErrors;
